requirements.txt
requests==2.32.3
typer==0.24.0
rich==13.9.4
pydantic==2.10.4
python-dotenv==1.0.1
pytest==8.3.4
pytest-mock==3.14.0
responses==0.25.6
